requirements.txt
pandas
urlextract
vaderSentiment
wordcloud
Pillow
emoji
streamlit
plotly
matplotlib
seaborn
